tests-pyodbc/requirements.txt
pyodbc>=5.0.1
pytest>=7.4.3
pyyaml
